Fix collision enable bug

+++ Engine/Stat.h
@@ -2,10 +2,10 @@
 
 struct Stat
 {
-	int32 hp = 100;
-	int32 maxHp = 100;
-	int32 attack = 20;
-	int32 defence = 5;
+	float hp = 100.f;
+	float maxHp = 100.f;
+	float attack = 20.f;
+	float defence = 5.f;
 	float speed = 0.f;
 	float maxSpeed = 100.f;
 };
